requirements.txt
absl-py>=0.9.0
albumentations>=0.4.5
astunparse>=1.6.3
cachetools>=4.1.1
certifi>=2020.6.20
chardet>=3.0.4
cycler>=0.10.0
decorator>=4.4.2
docopt>=0.6.2
future>=0.18.2
gast>=0.3.3
google-auth>=1.18.0
google-auth-oauthlib>=0.4.1
google-pasta>=0.2.0
grpcio>=1.30.0
h5py>=2.10.0
idna>=2.10
imageio>=2.9.0
imgaug>=0.2.6
importlib-metadata>=1.7.0
Keras-Preprocessing>=1.1.2
kiwisolver>=1.2.0
Markdown>=3.2.2
matplotlib>=3.2.2
networkx>=2.4
numpy>=1.19.0
oauthlib>=3.1.0
opencv-python>=4.3.0.36
opt-einsum>=3.2.1
Pillow>=7.2.0
protobuf>=3.12.2
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pyparsing>=2.4.7
python-dateutil>=2.8.1
PyWavelets>=1.1.1
PyYAML>=5.3.1
requests>=2.24.0
requests-oauthlib>=1.3.0
rsa>=4.6
scikit-image>=0.17.2
scipy>=1.4.1
six>=1.15.0
tensorboard>=2.2.2
tensorboard-plugin-wit>=1.7.0
tensorflow>=2.2.0
tensorflow-estimator>=2.2.0
termcolor>=1.1.0
tifffile>=2020.7.4
torch>=1.2.0
torchvision>=0.4.0
urllib3>=1.25.9
Werkzeug>=1.0.1
wrapt>=1.12.1
zipp>=3.1.0
